requirements.txt
beartype==0.15.0
black==23.3.0
diffusers==0.21.2
fastapi==0.96.0
facenet-pytorch==2.5.3
ftfy==6.1.3
langchain==0.0.262
loguru==0.7.0
numpy==1.24.1
openai==1.2.4
pillow==9.3.0
pretrainedmodels==0.7.4
regex==2023.6.3
sqlalchemy==2.0.19
starlette==0.27.0
streamlit==1.24.0
torch==2.0.1+cu117
torchaudio==2.0.2+cu1174
torchvision==0.15.2+cu117
tqdm==4.65.0
uvicorn==0.22.0
ujson==5.7.0
git+https://github.com/openai/CLIP.git
pymysql==1.4.6


playsound==1.2.2
SpeechRecognition==3.10.1
gtts==2.5.1
pyaudio==0.2.14
deep-translator==1.11.4